Removing Windows warnings

+++ source/lib/PccLibCommon/source/PCCContext.cpp
@@ -62,7 +62,7 @@ void PCCAtlasContext::resize( size_t size, size_t frameStart ) {
 }
 
 size_t PCCAtlasContext::calculateAFOCLsb( size_t frameOrder ) {
-  return frameOrder % ( 1 << log2MaxAtlasFrameOrderCntLsb_ );
+  return frameOrder % ( size_t( 1 ) << log2MaxAtlasFrameOrderCntLsb_ );
 }
 
 size_t PCCContext::calculateAFOCval( std::vector<AtlasTileLayerRbsp>& atglList, size_t atglOrder ) {
@@ -79,7 +79,7 @@ size_t PCCContext::calculateAFOCval( std::vector<AtlasTileLayerRbsp>& atglList, 
   auto&  afps                    = getAtlasFrameParameterSet( atgh.getAtlasFrameParameterSetId() );
   auto&  asps                    = getAtlasSequenceParameterSet( afps.getAtlasSequenceParameterSetId() );
 
-  size_t maxAtlasFrmOrderCntLsb  = 1 << ( asps.getLog2MaxAtlasFrameOrderCntLsbMinus4() + 1 );
+  size_t maxAtlasFrmOrderCntLsb  = size_t( 1 ) << ( asps.getLog2MaxAtlasFrameOrderCntLsbMinus4() + 1 );
   size_t afocLsb                 = atgh.getAtlasFrmOrderCntLsb();
   size_t prevAtlasFrmOrderCntLsb = atglList[atglOrder - 1].getHeader().getAtlasFrmOrderCntLsb();
   if ( ( afocLsb < prevAtlasFrmOrderCntLsb ) &&
